+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
   * API breaking release!
   * Rewritten to recurse into an array rather an object - see readme
+  * Now throws when given a recurse key not present on root
   * Test coverage and documentation improvements
 
 0.0.2 / 2012-10-20